+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+## Added
+
+- `tax`: added support for `informative` tax categories that will be calculated and reported but will not affect the invoice totals.
+- `br`: made ISS an informative tax
+
 ## [v0.220.6] - 2025-08-12
 
 ### Changed